// console/src/generated/ChannelSupportDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RoutingImplementationDto } from "./RoutingImplementationDto";

export type ChannelSupportDto = { source: string, target: string, operation: string, target_operation: string, group: string, implementation: RoutingImplementationDto, };