// console/src/generated/RoutingRuleWriteRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RoutingImplementationDto } from "./RoutingImplementationDto";

export type RoutingRuleWriteRequest = { provider_id: number, operation: string, kind: string, implementation: RoutingImplementationDto, dest_operation: string | null, dest_kind: string | null, sort_order: number, enabled: boolean, };